+++ src/Operator.java
@@ -1,22 +1,22 @@
 public enum Operator {
 
     ADDOPERATOR() {
-        public int operate(int a, int b) {
+        public double operate(int a, int b) {
             return a + b;
         }
     },
     SUBSTRACTOPERATOR {
-        public int operate(int a, int b) {
+        public double operate(int a, int b) {
             return a - b;
         }
     },
     MULTIPLYOPERATOR {
-        public int operate(int a, int b) {
+        public double operate(int a, int b) {
             return a * b;
         }
     },
     DIVIDEOPERATOR {
-        public int operate(int a, int b) {
+        public double operate(int a, int b) {
             if (b == 0) {
                 new DivisionByZeroException("0으로 나눌 수 없습니다.");
             }
@@ -24,7 +24,7 @@ public enum Operator {
         }
     },
     MODOPERATOR {
-        public int operate(int a, int b) {
+        public double operate(int a, int b) {
             if (b == 0) {
                 new DivisionByZeroException("0으로 나눌 수 없습니다.");
             }
@@ -32,5 +32,5 @@ public enum Operator {
         }
     };
 
-    public abstract int operate(int a, int b);
+    public abstract double operate(int a, int b);
 }
